backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-socketio==5.11.2
python-engineio==4.9.0
websockets==12.0
httpx==0.27.0
